+++ src/ui/tv_show/TvShowWidgetTvShow.h
@@ -64,13 +64,14 @@ private slots:
     void onArtPageTwo();
     void onDownloadTune();
 
-    void onNameChange(QString text);
+    void onTitleChange(QString text);
+    void onSortTitleChange(QString text);
+    void onOriginalTitleChange(QString text);
     void onImdbIdChange(QString text);
     void onTmdbIdChange(QString text);
     void onTvdbIdChange(QString text);
     void onTvMazeIdChange(QString text);
     void onStatusChange(int index);
-    void onSortTitleChange(QString text);
     void onCertificationChange(QString text);
     void onRatingChange(double value);
     void onUserRatingChange(double value);